tsm: add TestCache_CacheTestCommitLock
A test to check that the commit lock is working as expected. In particular,
we want to check that if there is a pending commit, a prepare issued
by a concurrent goroutine will block until such time as the
current holder of the lock either commits or rolls back.

This test also checks that concurrent writes into the cache
are not prevented by goroutines that hold the cache's commit lock.

Signed-off-by: Jon Seymour <jon@wildducktheories.com>