tsdb: introduce EnsureUniqueNonEmptyNames
This function is required to ensure that each field of a converted
Column has a unique, non-empty name.

Non empty names are required to ensure that the newly enforced
constraints on points.NewPoints are always satisfied, even if
arithmetical InfluxQL expressions are used (such columns have blank names).

Unique names are required so that we don't silently elide columns
that happen to have the same name.

Signed-off-by: Jon Seymour <jon@wildducktheories.com>