tsm: cache: improve thread safety of Cache.Deduplicate.
Fixes #5699.

Cache.Deduplicate is designed to be execute while the caller is not holding
any locks.

The entry.deduplicate() method contains two statements:

e.values = e.values.Deduplicate()
e.needSort = false

which need to be executed in that order (as observed by all threads). However,
without a synchronization construct won't be.

This fix makes a clone of each entry before performing the deduplication,
then while briefly holding the lock, updates the store with the deduplicated clones

A second fix guards access to snapshots with a readlock.

Also a small tweak to remove unnecessary deduping of single element lists.

Signed-off-by: Jon Seymour <jon@wildducktheories.com>