stats: allow statistics maps to quietly fade away when the object they refer to is closed
As per #5784, there is a risk that a long running influxd process will accumulate a lot of statistics
that describe objects that have already been closed.

This is a concern primarily because of the amount of IO and CPU involved with continuing to publish the last will and testament of an object long since departed.

This series address the issue by pinching some code from Kapacitor - thanks @nathanielc - and using
it as a top level name space for influx statistics objects. The maps are reference counted. One reference is allocated to the described object, another to the monitor.

When the described object is closed, it yells DeleteStatistics! at the influxdb package which takes note of the request, but otherwise ignores it. When the monitor is doing its thing, it quietly peeks at the statistics map's reference count and if it detects that its number is up (well, down really), it yells DeleteStatistics! at the influxdb package who takes notice of it this time. This deferral of the grim reaper is required so that the monitor has time to tidy up the paper work and confirm that the undertaker's cheque has not bounced.

Signed-off-by: Jon Seymour <jon@wildducktheories.com>