squash-me: introduce a dedicate StatisticsSet interface
The intent of these types is to provide a type safe way to manage the
lifecycle of an influx statistics set.

The `statistics` type has a 3-phase lifecycle: building, running and closed.

During the building phase, new fields can be added but the statistics are
not yet published. Once the MustBuild() method is called the runtime
statistics map is created and the receiver moves into the runtime state.
Further mutation of the `intVars`, `floatVars` and `stringVars` map is
not permitted in this state - any attempt to do so results in a panic.
This also means that access to the var maps can be done in a lock free
fashion, since the var maps can never be modified.

Once the described object is closed, it can call the Close() method
which causes the runtime to stop publishing the statistics.

Assuming this design is accepted, further commits will refactor these
types and `influxdb `so that the implementation of `influxdb.NewStatistics()`
is pulled up into `statistics.Build()` and then the implementation
of `influxdb.NewStatistics()` changed to make use of these types. Similar
changes will be made to simplify the implementation of the `monitor`
package.

Signed-off-by: Jon Seymour <jon@wildducktheories.com>