squash-me: stats: extensive refactoring
The intent of these types is to provide a type safe way to manage the
lifecycle of an influx statistics set.

The `statistics` type has a 3-phase lifecycle: building, running and closed.

During the building phase, new fields can be added but the statistics are
not yet published. Once the MustBuild() method is called the runtime
statistics map is created and the receiver moves into the runtime state.
Further mutation of the `intVars`, `floatVars` and `stringVars` map is
not permitted in this state - any attempt to do so results in a panic.
This also means that access to the var maps can be done in a lock free
fashion, since the var maps can never be modified.

Once the described object is closed, it can call the Close() method
which causes the runtime to stop publishing the statistics.

Assuming this design is accepted, further commits will refactor these
types and `influxdb `so that the implementation of `influxdb.NewStatistics()`
is pulled up into `statistics.Build()` and then the implementation
of `influxdb.NewStatistics()` changed to make use of these types. Similar
changes will be made to simplify the implementation of the `monitor`
package.

* rename DeleteStatistics to CloseStatistics
* rename StatisticsSet to Statistics
* rename StatisticsSetBuilder to Builder
* introduce a Regdistry interface which is the manages the set of all Statistics
* introduce a stats.Root var which implements Stats.Registry
* reimplement influxdb package in terms of stats package
* reimplement monitor package in terms of stats package
* introduce an Openable() interface to force a Builder product to be opened to be useful
* make reference counting an explicit part of StatisticsSet interface
* removed edited expvar package - no longer needed

Signed-off-by: Jon Seymour <jon@wildducktheories.com>