squash-me: arrange for the monitor to take a reference on each stats object.
Previously the reference counting had a little bit too much black magic - the
registry automatically assigned a non-zero reference count to each
newly created Statistics object and deemed that one of these was
for the site of usage, and one for the monitor. However, this built in the
assumption that there was a monitor and a site of usage.

Now the reference counter is updated by explicit Open() calls made by
the site of usage and the monitor. The monitor needs a way to know that
there is an Open call it needs to make. So, we rename Registry.Open() to
Registry.NotifyOpen() and provide a means to register a listener with the Registry.

Now, whenever a site of usage opens a newly constructed statistics object,
the object asks the registry to notify any listeners. These listeners
can choose to open the object if they wish to.

Signed-off-by: Jon Seymour <jon@wildducktheories.com>