stats: Introduce a new stats API.
The API has several goals:

* retain an underlying expvar implementation
* expose an easy to use API for runtime code to use
* allow statistics to be forgotten once the object described by the statistics has been closed
* avoid the need for a big bang conversion to the new API

The API uses a builder pattern which requires that described objects declare
the members of their own statistics set during a build phase before they are first
used at runtime. Once the build phase is complete the statistics object is explicitly "opened"
and then an interface is exposed that allows the runtime to set or add values to the
set of statistics pre-declared during the build phase.

This two phase approach allows us to reduce the amount of locking needed during
the runtime phase since, by construction, we prevent certain classes
of mutation during the runtime phase. We use the go type system to help
enforce correct construction of objects by preventing access to the
runtime interface until the object has first been defined, built and finally
opened.