squash-me: stats: add a TestMonitor
The test monitor mimics the essential behaviour of the monitor for the
purposes of test cases. In particular, it registers an observer for
new Statistics objects, acquires a reference upon notification and
releases its own reference once it detects that the number of references
has dropped to one.

The output of its Observe() method can used to observe the set of Statistics
the monitor would see if it polled the Registry at a particular point in time.

Signed-off-by: Jon Seymour <jon@wildducktheories.com>