stats: add support for measuring the idle time of a statistics object.
Long lived, but otherwise idle, objects can generate a lot of uninteresting
statistics data. The overheads of doing are not insignificant.

This change keeps track of how often the statistics are being updated use
the concept of a busy counter, updated by the recorder, and idleSince
timestamp, updated by interested views.

Each Statistics reports the number of nanoseconds since it was last updated.
If this value exceeds a certain amount, then the monitor skips the
statistics when copying the statistics into the database.

We need to maintain support for statistics that either have not adopted
the new API or for which this kind of suppression is not desired. To support
these statistics, we add the DisableIdleTimer() support to the Builder
interface.

We also need to be able to ignore certain statistics where a change
to that statistic by itself, is not interesting. For these. we
add a method call DontUpdateBusyCount() which prevents an otherwise
uninteresting update from keeping an otherwise idle statistic busy.

Signed-off-by: Jon Seymour <jon@wildducktheories.com>