stats: monitor: implement support for an idle time on statistics.
We update the monitor to check whether a Statistics object is idle
before doing any work to write it into a database.

The monitor configuration is updated to include a new member 'idle-time'
which is a toml.Duration which is the time after which an idle
Statistics object is ignored. The default timeout is 1 hour.

We change influxdb.NewStatistics() to ensure that any Statistics object that
is updated via the legacy expvar interface does not have its
reporting function disabled.

Signed-off-by: Jon Seymour <jon@wildducktheories.com>