squash-me: stats: add IsOpen() to the recorder interface.
It has been discovered that some described objects can be repeatedly
closed once already closed or can fail while being opened only to
be opened a second time without an intervening close.

One example is DeleteShard() which will be repeatedly called if
it fails and performs a Shard.Close() operation on each attempt.

Rather than panic because the Statistics object has already been
opened or closed we need to more gracefully handle this case
so that root cause isn't masked by the symptom (that the Statistics
object has been closed or opened twice in a row).