squash-me: stats: tsdb: tolerate Open() or Close() being called twice in row
In certain error conditions, Shard.Open() or Shard.Close() can fail. When
this happens, the caller may attempt to retry the operation a second
time. This operation won't necessarily succeed, but if stats.Recorder
blows up, it can obscure the error condition that led to the consecutive
Open() or Close() calls being made.

To allow described objects to have idempotent Open() or Close() methods
if they so choose, or at least, not prevent them doing so we add a guard
to each call to stats.Recorder.Open() and stats.Recorder.Close() to
check that the call will not fail because of the Recorder's existing
state.

Tha main reason for doing this is to ensure that the panic that
would otherwise result doesn't obsecure error messages that
might otherwise be logged which would explain why the consecutive
Open() or Close() methods on the containing described object were called.

Signed-off-by: Jon Seymour <jon@wildducktheories.com>