improve feedback from admin UI regarding submission of queries.
It can sometimes be difficult to determine if a query submitted by
the UI has been accepted for execution.

In particular, if a query that returns empty results is followed
by a query that takes a long time, then the green success message
from the first query may be misleadingly displayed to the user when,
in fact, the second query is still in progress.

To address this, we always hide the query error and success divs
before the query is submitted. Previously, just the results were
cleared.

Secondly, if the user re-runs a query expecting slightly different results,
it can also be unclear whether the second attempt to execute the command
is simply redisplaying the results of the previous query or the results of the
resubmission, this is particularly true if the queries involved
have short execution times.

To support the ability to distinguish these two cases, we have any attempt
to use the history arrow also clear the results and status divs. This
reduces the ambiguity about whether the next results display is, in fact,
the result of executing a new query.

Signed-off-by: Jon Seymour <jon@wildducktheories.com>